+++ src/XRoadLib.Extensions.AspNetCore/WebServiceContextAccessor.cs
@@ -1,0 +1,12 @@
+namespace XRoadLib.Extensions.AspNetCore
+{
+    public interface IWebServiceContextAccessor
+    {
+        WebServiceContext WebServiceContext { get; }
+    }
+    
+    public class WebServiceContextAccessor : IWebServiceContextAccessor
+    {
+        public WebServiceContext WebServiceContext { get; set; }
+    }
+}